Wrapper for snpdiff
Wrapper removes the need for a user to rename the fields in the column
header of a genotype file, and to convert a 1-based set of positions to
0-based. User passes the names or numbers of the fields to be changed to
match snpdiff input requirements and the offset if necessary. Writes a
temporary file which conforms to snpdiffs expectations and passes the
file to snpdiff.